requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
beautifulsoup4==4.13.4
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
Crawl4AI==0.6.2
cryptography==44.0.3
cssselect==1.3.0
distro==1.9.0
exceptiongroup==1.2.2
fake-http-header==0.3.5
fake-useragent==2.2.0
fastmcp==2.3.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
google-auth==2.40.1
google-genai==1.14.0
greenlet==3.2.1
grpcio==1.67.1
h11==0.16.0
h2==4.2.0
hf-xet==1.1.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.1
humanize==4.12.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
litellm==1.68.1
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.8.0
mdurl==0.1.2
milvus-lite==2.4.12
multidict==6.4.3
nltk==3.9.1
numpy==2.2.5
ollama==0.4.8
openai==1.78.0
openapi-pydantic==0.5.1
packaging==25.0
pandas==2.2.3
pillow==10.4.0
pip-autoremove==0.10.0
playwright==1.52.0
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.1
pymilvus==2.5.8
pyOpenSSL==25.0.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
setuptools==80.3.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.7
sse-starlette==2.3.4
starlette==0.46.2
tf-playwright-stealth==1.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.2
websockets==14.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
